Feedback on Lab 8

+++ Lab/Lab8/problem11_17.java
@@ -15,7 +15,7 @@ public class problem11_17 {
 		int smallestNumber = 1;
 		
 		for(int i=0; i<integerFactors.size();i++) {
-			smallestNumber = smallestNumber * integerFactors.get(i);
+			smallestNumber = smallestNumber * (int)integerFactors.get(i);
 		}
 		System.out.println("The smallest number n for m *n to be a perfect square is "+ smallestNumber);
 		System.out.println("m * n is " +(m*smallestNumber));
@@ -32,10 +32,11 @@ public class problem11_17 {
 			}
 		}
 
+    // JA: This does not work
 	public static void getOddNumberedFactors(ArrayList list) {
 		for (int i=0; i<list.size()-1; i++) {
 			for(int j=i+1; j<list.size(); j++) {
-				if(list.get(i) == list.get(j)) {
+				if(list.get(i) == list.get(j)) { 
 					list.remove(j);
 					list.remove(i);
 				}
